osx_requirements.txt
-e git+https://github.com/CoreSecurity/pcapy.git@2984bf5e974ab3a2af58c0501f2d4072db5ac4e4#egg=pcapy-10.8
dnet
dpkt==1.9.2
hexdump
mock
nose
psutil==5.6.1
scapy==2.4.2
six==1.12.0
twitter.common.log==0.3.11
